Moved page tests
