+++ src/tests/Pagetests/aboutuspage.test.tsx
@@ -1,5 +1,5 @@
 import { render } from "@testing-library/react";
-import AboutUsPage from "../pages/aboutuspage";
+import AboutUsPage from "../../pages/aboutuspage";
 
 test("Check if page renders", () => {
   const page = render(<AboutUsPage />);
