+++ src/tests/Pagetests/artspage.test.tsx
@@ -1,5 +1,5 @@
 import { render } from "@testing-library/react";
-import ArtesPage from "../pages/artspage";
+import ArtesPage from "../../pages/artspage";
 
 test("Check if page renders", () => {
   const page = render(<ArtesPage />);
