+++ src/tests/Pagetests/landingpage.test.tsx
@@ -1,5 +1,5 @@
 import { render } from "@testing-library/react";
-import LandingPage from "../pages/landingpage";
+import LandingPage from "../../pages/landingpage";
 
 test("Check if page renders", () => {
   const page = render(<LandingPage />);
